# inject_panic_runtime(): Avoid double negation for 'any non rlib'