@lcnr 2024-11-22
r? compiler