@Zalathar 2024-11-23
I'm not very familiar with the injection of `profiler_builtins`, but based on my limited understanding I suspect that this might be the wrong fix.

I did some playing around, and this might potentially be the more appropriate solution:
- Change profiler_builtins from `#[no_std]` to `#[no_core]`, since it doesn't contain any actual Rust code.
- Remove the check that refuses to inject profiler_builtins into `#[no_core]` crates.