@Zalathar 2024-11-23
Basically, I think the underlying problem is that because profiler_builtins is currently `#[no_std]`, it still has an implicit dependency on core. That causes problems when trying to make it a dependency of core itself.

But I'm pretty sure that:
- There is no need for profiler_builtins to actually depend on core.
- When built with `-Cinstrument-coverage`, core really should have a genuine dependency on profiler_builtins, to make sure the profiler runtime ends up in the final binary even if no other crate in the final binary is built with instrument-coverage.