@Zalathar 2024-11-23
For reference, the error message was added in #79958. It looks like the real issue (profiler doesn't need core) wasn't noticed at that time.