@Enselic 2024-11-23
Your solution is clearly better than mine, and I have confirmed it works for me too. Thanks!

I have now made this PR only fix the double-negation that bjorn3 pointed out.